Move cnl::digits into a separate header file
